[REFACTOR]: Refactor components and remove unused code
