+++ components/FloatingButton/index.tsx
@@ -27,8 +27,8 @@ export default function FloatingButton() {
   const handleCloseCreateNoti = () => setOpenedCreateNoti(false);
 
   return (
-    <div className='group fixed bottom-16 right-10 flex h-24 w-24 items-end justify-end p-2 *:cursor-pointer'>
-      <div className='flex-center absolute z-50 rounded-full bg-gradient-to-r from-blue-800 to-violet-800 p-3 text-white shadow-xl'>
+    <div className='group fixed bottom-5 right-24 flex h-24 w-24 items-end justify-end p-2 *:cursor-pointer'>
+      <div className='flex-center absolute z-50 rounded-full bg-gradient-to-r from-blue-600 to-violet-600 p-3 text-white shadow-xl'>
         <FaPlus className='size-6' />
       </div>
