+++ components/ToTopButton/index.tsx
@@ -10,15 +10,15 @@ export default function ToTopButton({ viewportRef }: ToTopButtonProps) {
   const { scrollY } = useStore((store) => store);
 
   return (
-    <Affix position={{ bottom: 20, right: 20 }}>
+    <Affix position={{ bottom: 30, right: 50 }}>
       <Transition transition='slide-up' mounted={scrollY > 100}>
         {(transitionStyles) => (
-          <Button
-            leftSection={<FaArrowUp style={{ width: rem(16), height: rem(16) }} />}
+          <div
+            className='cursor-pointer rounded-full bg-violet-600 p-3 shadow-lg'
             style={transitionStyles}
             onClick={() => viewportRef.current!.scrollTo({ top: 0, behavior: 'smooth' })}>
-            Scroll to top
-          </Button>
+            <FaArrowUp style={{ width: rem(16), height: rem(16) }} />
+          </div>
         )}
       </Transition>
     </Affix>
